requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
drf-yasg>=1.21
psycopg2-binary>=2.9
Pillow>=10.0
celery>=5.5.2
redis>=5.0

